[model, worker, cfg] fix: stabilize ACT SAC training
Separate rollout exploration noise from deterministic actor training, correct ACT batching and gradient boundaries, and persist newly collected replay data.\n\nUpdate the maintained launcher with the validated fixed-replay TD3+BC recipe.

+++ src/verl_vla/models/act_torch/policy/base.py
@@ -41,7 +41,7 @@ class ActOutput:
     def from_model_output(cls, model_output) -> "ActOutput": ...
 
     def to_data_proto(self) -> DataProto:
-        tensor_batch = {"action": self.action}
+        tensor_batch = {"action": self.action.float()}
         if self.log_prob is not None:
             tensor_batch["log_prob"] = self.log_prob
         return DataProto.from_dict(tensors=tensor_batch)
